Add a new messaging-clustered module (#111)
This is a first pass at getting basic messaging clustering
working. Right now lots of things are hardcoded, including the cluster
password, so it's not yet suitable for production usage. Once we get
the config-api working with messaging then all these values can be
configured.

To enable clustering, just use wildfly-swarm-messaging-clustered as
a dependency instead of wildfly-swarm-messaging.

As part of this change, I've added a priority with a default value of
0 to ServerConfiguration. This priority is used when more than one
ServerConfiguration implementation have the same Fraction type T. The
ServerConfiguration with the higher priority is the only one that gets
used.