Delegate all static resource handling to Undertow itself
For hot reloading of changes during development, we now write out a
WEB-INF/undertow-external-mounts.conf. A side-effect of this is that
static content is now always served out of the same context as your
application. If static content needs to be served at a different
context than your application then that static content needs to reside
in a subdirectory of src/main/resources or src/main/webapp.