Rework logging.
Initially, logging for BootstrapLoggers is configured via properties:

  -Dswarm.log.CATEGORY.NAME=LEVEL

Once JBoss-Logging is installed, those same properties are transferred and merged
with the logging.properties and subsequently through to the logging subsystem configuration
to allow DMR-management of bootstrap loggers.

Additionally, while the bootstrap loggers are initially backed by a simple (needs improvement)
console logger, once JBoss-Logging is installed, they are re-attached to a JBL Logger.