Alter how foo.jar and foo-modules.jar (and other .jars that contain modules) are bootstrapped:
- If a jar contains modules, it should include `wildfly-swarm-modules.conf`
  as an empty file

- A fraction that needs to be bootstrapped should include `wildfly-swarm-bootstrap.conf`
  containing the name(s) of module(s) to bootstrap.

Basically, all wildfly-swarm-bootstrap.conf files that lived in a modules/ project
which had a peer api/ project have been moved to the api/ project, and replaced
with an empty wildfly-swarm-moduules.conf file in the modules/ project.

For any project (such as guava or bean-validation), the -bootstrap.conf file has been
replaced entirely with the -modules.conf file.

Bottom line:

  - wildfly-swarm-modules.conf signals that the .jar contains a modules/ hierarchy
  - wildfly-swarm-bootstrap.conf signals that the .jar contains a bootstrappable
    fraction.