Provide a swagger-webapp fraction.
Motivation:
We already provide a swagger fraction that allows developers to expose
their APIs via /swagger.json. The `swagger-webapp` fraction provides a
simple way to expose the Swagger UI http://swagger.io/swagger-ui/ as
well.

The UI may be run as its own standalone deployment, or it may be
included as a part of another deployment.

Modifications:
Added the `swagger-webapp` fraction. Adjusted /pom.xml and the bom
accordingly.

Result:
Currently the UI seems to be working in the simple InVM test, but not in
the Arquillian test. This commit is a checkpoint.