SWARM-278: Notify onTopologyChange listeners immediately
Make sure that when an onTopologyChange listener is registered in
topology.js that we immediately invoke it with a snapshot of the
current topology. Otherwise, listeners may miss the initial topology
change event that gets fired when first connecting to the SSE stream.

The object passed to the listener is now the parsed JSON object
instead of the JSON string for consistency.