Make swagger-webapp implicit & allow customization.
Motivation:
Previously, the swaggwer-webapp fraction overlaid its content on to the
user's deployment. In most cases this is fine, because swagger-webapp
typically would stand alone as a separate process. However, we don't
want to prevent users from including swagger-webapp in some existing
application.

Additionally, typical instances of swagger-ui in the real world include
some customization of the UI itself through HTML, CSS and Javascript
artifacts. We need to support this scenario.

Modifications:
The `SwaggerWebAppConfiguration` class now deploys the swagger-ui bits
as an implicit deployment alongside the user's deployment, but still in
the same container. They `SwaggerWebAppFraction` class has been modified
to support user-provided content which is overlaid onto the swagger-ui
bits. This content can be provided as a path to a directory or a `jar`|`war`
file on disk, or as a GAV with maven coordinates.

Result:
Developers using swagger-webapp can now customize the resulting user
interface, and don't have to worry about our crap messing with their
crap or vice versa.