Don't include the app artifact when running from the classpath [SWARM-218]
Motivation:
Depending on how the an app is packaged/started, there is an
inconsistency with the app artifact when calling addAllDependencies().
For -swarm.jars, it's not included in the dependencies, but for the run
task, IDE run, or Arquillian, the app artifact is on the classpath, so
is included by addAllDependencies().

Modifications:
* Filter out all classpath elements that are in the current directory,
  not just directories

Result:
The app artifact is no longer included in the set of all dependencies.