Provide public access to the default deployment [SWARM-286]
Motivation:
Applications with custom mains currently have to try to recreate some or
all of the default deployment. If we make it available, they can augment
it, or add it as a library (in the case of wars) to the archive they are
building.

Modifications:
* expose Container#createDefaultDeployment()
* remove unused Container from the deployment factory interface

Result:
Users now have access to the default deployment.